Fix 2 unit tests for 2.1.0

+++ src/main/java/com/fasterxml/jackson/dataformat/csv/CsvSchema.java
@@ -290,7 +290,7 @@ public class CsvSchema
          * Default is "not used", meaning that no escaping used.
          */
         public Builder setEscapeChar(char c) {
-            _escapeChar = (int) c;
+            _escapeChar = c;
             return this;
         }
